scripts/requirements.txt
python-dotenv>=1.0.0
supabase>=2.0.0
twilio>=9.0.0
pytz>=2024.1

